+++ test/integration/060_persist_docs_tests/test_persist_docs.py
@@ -284,3 +284,38 @@ class TestPersistDocsNested(BasePersistDocsTest):
 
             level_3_column = node['columns']['level_1.level_2.level_3_a']
             assert level_3_column['comment'] == "level_3 column description"
+
+
+class TestPersistDocsColumnMissing(BasePersistDocsTest):
+    @property
+    def project_config(self):
+        return {
+            'config-version': 2,
+            'models': {
+                'test': {
+                    '+persist_docs': {
+                        "columns": True,
+                    },
+                }
+            }
+        }
+
+    @property
+    def models(self):
+        return 'models-column-missing'
+
+    @use_profile('snowflake')
+    def test_snowflake_missing_column(self):
+        self.run_dbt()
+        self.run_dbt(['docs', 'generate'])
+        with open('target/catalog.json') as fp:
+            catalog_data = json.load(fp)
+        assert 'nodes' in catalog_data
+
+        table_node = catalog_data['nodes']['model.test.missing_column']
+        table_id_comment = table_node['columns']['ID']['comment']
+        assert table_id_comment.startswith('test id column description')
+
+    @use_profile('bigquery')
+    def test_bigquery_missing_column(self):
+        self.run_dbt()